backend/requirements.txt
debugpy==1.8.6
fastapi==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.35
psycopg2==2.9.9 
websockets==13.1
pillow==10.4.0
